requirements.txt
requests==2.31.0
Pillow==10.2.0
telethon==1.34.0
python-dotenv==1.0.1
rich==13.7.1
